requirements.txt
Django==1.5
django-allauth==0.18.0
python-instagram==1.3.0
ipdb==0.8
